backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pillow>=10.0.0
mediapipe>=0.10.14
numpy>=1.24